+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [unreleased]
+
+### Fixed
+- Fixed editor crash by changing Editor.js and plugins to static UMD versions, preventing breaking changes from `@latest` CDN releases (specifically `@editorjs/paragraph` version 3.0.0).
+
 ## [0.2.3] - 2026-07-07
 
 ### Added